[SearchPlace] Modify architecture by ViewModel
 - Create SearchPlaceViewModel
 - Fix tmap api init in ViewModel
 - Modify place search in ViewModel
